Separate cobra & K8s related code into diff packages
