+++ cmd/get/volume.go
@@ -1,0 +1,30 @@
+package get
+
+import (
+	"github.com/openebs/openebsctl/pkg/cstor/get"
+	"github.com/openebs/openebsctl/pkg/util"
+	"github.com/spf13/cobra"
+)
+
+var (
+	volumesListCommandHelpText = `
+This command displays status of available zfs Volumes.
+If no volume ID is given, a list of all known volumes will be displayed.
+
+Usage: kubectl openebs cStor volume list [options]
+`
+)
+
+// NewCmdGetVolume displays status of OpenEBS Volume(s)
+func NewCmdGetVolume() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "volume",
+		Aliases: []string{"vol", "v", "volumes"},
+		Short:   "Displays status information about Volume(s)",
+		Long:    volumesListCommandHelpText,
+		Run: func(cmd *cobra.Command, args []string) {
+			util.CheckErr(get.RunVolumesList(cmd), util.Fatal)
+		},
+	}
+	return cmd
+}
