+++ pkg/cstor/describe/pvc_info.go
@@ -30,18 +30,6 @@ import (
 	"k8s.io/cli-runtime/pkg/printers"
 )
 
-var (
-	pvcInfoCommandHelpText = `
-This command fetches information and status  of  the  various  aspects 
-of  the  PersistentVolumeClaims  and  its underlying related resources 
-in the provided namespace. If no namespace is provided it uses default
-namespace for execution.
-
-$ kubectl openebs describe pvc [name1] [name2] ... [nameN] -n [namespace]
-
-`
-)
-
 const (
 	cstorPvcInfoTemplate = `
 {{.Name}} Details :
@@ -81,26 +69,6 @@ SIZE             : {{.Size}}
 PV STATUS    	 : {{.PVStatus}}
 `
 )
-
-// NewCmdDescribePVC Displays the pvc describe details
-func NewCmdDescribePVC() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "pvc",
-		Aliases: []string{"pvcs", "persistentvolumeclaims", "persistentvolumeclaim"},
-		Short:   "Displays PersistentVolumeClaim information",
-		Long:    pvcInfoCommandHelpText,
-		Example: `kubectl openebs describe pvc cstor-vol-1 cstor-vol-2 -n storage`,
-		Run: func(cmd *cobra.Command, args []string) {
-			var pvNs, openebsNamespace string
-			if pvNs, _ = cmd.Flags().GetString("namespace"); pvNs == "" {
-				pvNs = "default"
-			}
-			openebsNamespace, _ = cmd.Flags().GetString("openebs-namespace")
-			util.CheckErr(RunPVCInfo(cmd, args, pvNs, openebsNamespace), util.Fatal)
-		},
-	}
-	return cmd
-}
 
 // RunPVCInfo runs info command and make call to display the results
 func RunPVCInfo(cmd *cobra.Command, pvcs []string, ns string, openebsNs string) error {
